feat: frame conversion improvements (#21)
- Preserve int32/int64 values as int64 columns instead of coercing every
  number to float64; a column that mixes ints and floats across rows still
  promotes to float64, the only one of the two that can represent both.
- Add a per-query flatten-depth option that keeps deeply nested documents
  beyond the configured depth as a single raw JSON column instead of fully
  exploding them into dot-notation columns.
- Add an explicit "long" format that returns a time-sorted long frame
  without attempting LongToWide, so callers who want long rows don't have
  to rely on that conversion's failure-and-fallback path.

+++ pkg/plugin/integration_test.go
@@ -182,7 +182,7 @@ func TestIntegrationQueryHandlers(t *testing.T) {
 		}
 		field := res.Frames[0].Fields[0]
 		v, ok := field.ConcreteAt(0)
-		if !ok || v.(float64) != 3 {
+		if !ok || v.(int64) != 3 {
 			t.Fatalf("count = %v, want 3", v)
 		}
 	})
